KIJIHIVE-54. Ensure state of current row is clear before calling pager methods.
When paging through a set of qualifiers on rows, if a row doesn't have data for
a given qualified column, an exception would be raised because a pager method
was called on a closed pager. To fix this, ensure that once a row's pagers are
closed, the row object is null.

Review: https://review.kiji.org/r/2071/
Reviewer: lsheng, adam